chore: update dependencies and extract auth action state
- Upgrade workspace dependencies and regenerate the lockfile
- Move shared auth action state into its own module
