+++ apps/api/package.json
@@ -15,28 +15,28 @@
   },
   "dependencies": {
     "@repo/logger": "workspace:*",
-    "body-parser": "^2.2.0",
-    "cors": "^2.8.5",
-    "express": "^4.21.2",
-    "morgan": "^1.10.0"
+    "body-parser": "^2.3.0",
+    "cors": "^2.8.6",
+    "express": "^4.22.2",
+    "morgan": "^1.11.0"
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
     "@repo/eslint-config": "workspace:*",
     "@repo/jest-presets": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@types/body-parser": "^1.19.5",
-    "@types/cors": "^2.8.17",
-    "@types/express": "~4",
-    "@types/morgan": "^1.9.9",
-    "@types/node": "^20",
-    "@types/supertest": "^6.0.2",
-    "esbuild": "^0.25.1",
+    "@types/body-parser": "^1.19.6",
+    "@types/cors": "^2.8.19",
+    "@types/express": "~4.17.25",
+    "@types/morgan": "^1.9.10",
+    "@types/node": "^20.19.43",
+    "@types/supertest": "^6.0.3",
+    "esbuild": "^0.25.12",
     "esbuild-register": "^3.6.0",
-    "eslint": "~9.39.2",
+    "eslint": "~9.39.5",
     "jest": "^29.7.0",
-    "nodemon": "^3.1.0",
-    "supertest": "^6.3.3",
+    "nodemon": "^3.1.14",
+    "supertest": "^6.3.4",
     "typescript": "5.9.3"
   }
 }
