+++ apps/react-router-web/package.json
@@ -21,7 +21,7 @@
     "clean": "rm -rf .react-router build"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.2.2",
+    "@hookform/resolvers": "^5.5.7",
     "@nichtsam/helmet": "0.3.0",
     "@react-router/express": "7.18.1",
     "@react-router/fs-routes": "7.18.1",
@@ -31,51 +31,52 @@
     "@sentry/profiling-node": "9.15.0",
     "@sentry/react-router": "9.15.0",
     "@supabase/ssr": "^0.6.1",
-    "@supabase/supabase-js": "^2.50.4",
+    "@supabase/supabase-js": "2.50.4",
     "@tusbar/cache-control": "1.0.2",
     "address": "^2.0.3",
-    "close-with-grace": "^2.2.0",
-    "compression": "^1.8.0",
-    "dotenv": "^17.1.0",
-    "execa": "^9.6.0",
-    "express": "^4.21.2",
+    "close-with-grace": "^2.5.0",
+    "compression": "^1.8.1",
+    "dotenv": "^17.4.2",
+    "execa": "^9.6.1",
+    "express": "^4.22.2",
     "express-rate-limit": "^7.5.1",
-    "get-port": "^7.1.0",
-    "isbot": "^5.1.28",
+    "get-port": "^7.2.0",
+    "isbot": "^5.2.1",
     "lucide-react": "^0.475.0",
-    "morgan": "^1.10.0",
+    "morgan": "^1.11.0",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
     "react-router": "7.18.1",
-    "remix-hook-form": "^7.1.1",
+    "react-hook-form": "7.71.1",
+    "remix-hook-form": "7.1.1",
     "remix-themes": "^2.0.4",
-    "sonner": "^2.0.5",
+    "sonner": "^2.0.7",
     "source-map-support": "^0.5.21",
     "spin-delay": "^2.0.1",
     "tiny-invariant": "^1.3.3",
-    "zod": "^4.3.6"
+    "zod": "^4.4.3"
   },
   "devDependencies": {
+    "@playwright/test": "^1.62.1",
     "@react-router/dev": "7.18.1",
-    "@playwright/test": "^1.55.0",
     "@repo/eslint-config": "workspace:*",
     "@repo/types": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@tailwindcss/vite": "^4.1.8",
+    "@tailwindcss/vite": "^4.3.3",
     "@total-typescript/ts-reset": "^0.6.1",
     "@types/compression": "^1.8.1",
-    "@types/express": "~4.17.21",
-    "@types/morgan": "^1.9.9",
-    "@types/node": "^20",
-    "@types/react": "^19.2.0",
-    "@types/react-dom": "^19.2.0",
+    "@types/express": "~4.17.25",
+    "@types/morgan": "^1.9.10",
+    "@types/node": "^20.19.43",
+    "@types/react": "^19.2.18",
+    "@types/react-dom": "^19.2.4",
     "@types/source-map-support": "^0.5.10",
     "@vitest/ui": "3.2.7",
-    "eslint": "~9.39.2",
-    "react-router-devtools": "^5.0.6",
-    "tailwindcss": "^4.1.18",
+    "eslint": "~9.39.5",
+    "react-router-devtools": "^5.1.6",
+    "tailwindcss": "^4.3.3",
     "typescript": "5.9.3",
-    "vite": "^6.3.5",
+    "vite": "^6.4.3",
     "vite-env-only": "^3.0.3",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "3.2.7"
