+++ apps/web/app/auth/action-state.ts
@@ -1,0 +1,6 @@
+export type AuthActionState = {
+  error?: string;
+  success?: string;
+};
+
+export const initialAuthActionState: AuthActionState = {};
