+++ apps/web/components/forgot-password-form.tsx
@@ -1,9 +1,7 @@
 "use client";
 
-import {
-  initialAuthActionState,
-  requestPasswordReset,
-} from "@/app/auth/actions";
+import { initialAuthActionState } from "@/app/auth/action-state";
+import { requestPasswordReset } from "@/app/auth/actions";
 import { cn } from "@repo/ui/lib/utils";
 import {
   Card,
