+++ apps/web/package.json
@@ -16,45 +16,45 @@
     "react-scan": "react-scan localhost:3000"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.2.2",
-    "@radix-ui/react-slot": "^1.2.3",
+    "@hookform/resolvers": "^5.5.7",
+    "@radix-ui/react-slot": "^1.3.3",
     "@repo/ui": "workspace:*",
     "@supabase/ssr": "^0.8.0",
-    "@supabase/supabase-js": "^2.89.0",
-    "@tanstack/react-query": "^5.83.0",
-    "@tanstack/react-table": "^8.21.2",
+    "@supabase/supabase-js": "^2.111.0",
+    "@tanstack/react-query": "^5.101.4",
+    "@tanstack/react-table": "^8.21.3",
     "class-variance-authority": "^0.7.1",
     "lucide-react": "^0.476.0",
-    "motion": "^11.17.0",
-    "next": "^16.2.11",
+    "motion": "^11.18.2",
+    "next": "16.2.11",
     "next-themes": "^0.4.6",
     "react": "^19.2.8",
     "react-day-picker": "9.10.0",
     "react-dom": "^19.2.8",
-    "react-hook-form": "^7.68.0",
-    "sonner": "^2.0.5",
+    "react-hook-form": "7.71.1",
+    "sonner": "^2.0.7",
     "zod": "^3.25.76"
   },
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
     "@repo/types": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@tailwindcss/postcss": "^4.1.18",
-    "@tanstack/eslint-plugin-query": "^5.81.2",
-    "@tanstack/react-query-devtools": "^5.83.0",
+    "@tailwindcss/postcss": "^4.3.3",
+    "@tanstack/eslint-plugin-query": "^5.101.4",
+    "@tanstack/react-query-devtools": "^5.101.4",
     "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
+    "@testing-library/react": "^16.3.2",
     "@total-typescript/ts-reset": "^0.6.1",
-    "@types/node": "^20",
-    "@types/react": "^19.2.0",
-    "@types/react-dom": "^19.2.0",
-    "@vitejs/plugin-react": "^5.1.1",
+    "@types/node": "^20.19.43",
+    "@types/react": "^19.2.18",
+    "@types/react-dom": "^19.2.4",
+    "@vitejs/plugin-react": "^5.2.0",
     "@vitest/coverage-v8": "4.1.10",
     "@vitest/ui": "4.1.10",
-    "eslint": "~9.39.2",
-    "jsdom": "^27.3.0",
+    "eslint": "~9.39.5",
+    "jsdom": "^27.4.0",
     "react-scan": "latest",
-    "tailwindcss": "^4.1.18",
+    "tailwindcss": "^4.3.3",
     "typescript": "5.9.3",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "4.1.10"
