+++ package.json
@@ -25,13 +25,13 @@
     "@changesets/cli": "^2.31.1",
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@types/node": "^20",
+    "@types/node": "^20.19.43",
     "@vitest/coverage-v8": "4.1.10",
-    "knip": "^6.2.0",
-    "prettier": "^3.5.3",
-    "prettier-plugin-jsdoc": "^1.3.2",
-    "prettier-plugin-tailwindcss": "^0.6.11",
-    "turbo": "^2.10.6",
+    "knip": "^6.29.0",
+    "prettier": "^3.9.6",
+    "prettier-plugin-jsdoc": "^1.8.1",
+    "prettier-plugin-tailwindcss": "^0.6.14",
+    "turbo": "^2.10.7",
     "typescript": "5.9.3",
     "vitest": "4.1.10"
   },
