+++ packages/eslint-config/package.json
@@ -9,17 +9,17 @@
     "./react-internal": "./react-internal.js"
   },
   "devDependencies": {
-    "@eslint/js": "~9.39.2",
-    "@next/eslint-plugin-next": "^16.2.11",
+    "@eslint/js": "~9.39.5",
+    "@next/eslint-plugin-next": "16.2.11",
     "@typescript-eslint/eslint-plugin": "^8.65.0",
     "@typescript-eslint/parser": "^8.65.0",
-    "eslint": "~9.39.2",
+    "eslint": "~9.39.5",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint-plugin-only-warn": "^1.2.1",
     "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^7.0.1",
-    "eslint-plugin-turbo": "^2.8.4",
-    "globals": "^17.3.0",
+    "eslint-plugin-react-hooks": "7.0.1",
+    "eslint-plugin-turbo": "^2.10.7",
+    "globals": "^17.8.0",
     "typescript": "5.9.3",
     "typescript-eslint": "^8.65.0"
   }
