+++ packages/jest-presets/package.json
@@ -7,6 +7,6 @@
   ],
   "dependencies": {
     "jest": "^29.7.0",
-    "ts-jest": "^29.3.0"
+    "ts-jest": "^29.4.12"
   }
 }
