+++ packages/logger/package.json
@@ -22,8 +22,8 @@
     "@repo/eslint-config": "workspace:*",
     "@repo/jest-presets": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@types/node": "^20",
-    "eslint": "~9.39.2",
+    "@types/node": "^20.19.43",
+    "eslint": "~9.39.5",
     "jest": "^29.7.0",
     "typescript": "5.9.3"
   }
